requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Migrations
alembic==1.13.1

# Environment
python-dotenv==1.0.0

# Forms and Validation
WTForms==3.1.1
Flask-WTF==1.2.1

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
